examples/email-template-renderer/requirements.txt
praisonai-svc>=1.2.0
python-dotenv>=1.0.0
Jinja2>=3.1.2



